// frontend/src/contracts/generated/contracts.ts
export const SAFE_BANK_DEPLOYMENT = {
  "network": "sepolia",
  "chainId": 11155111,
  "contracts": {
    "MockUSDC": {
      "address": "0xcf779EC5D80573D3254054a17c5B4f0117491662",
      "deploymentBlockNumber": 11320265,
      "explorerUrl": "https://sepolia.etherscan.io/address/0xcf779EC5D80573D3254054a17c5B4f0117491662#code",
      "verified": true
    },
    "VaultManager": {
      "address": "0xA79F660FaB4Ebae6Ac4298034Cb3FD6d28e5D2f7",
      "deploymentBlockNumber": 11320267,
      "explorerUrl": "https://sepolia.etherscan.io/address/0xA79F660FaB4Ebae6Ac4298034Cb3FD6d28e5D2f7#code",
      "verified": true
    },
    "SavingCore": {
      "address": "0xa35c55e7E2dB5874699cC9fb8d0E25032f51b443",
      "deploymentBlockNumber": 11320269,
      "explorerUrl": "https://sepolia.etherscan.io/address/0xa35c55e7E2dB5874699cC9fb8d0E25032f51b443#code",
      "verified": true
    }
  }
} as const;